Move the shard prefetch into distributed/utils.py
